requirements.txt
alabaster==0.7.12
allennlp==0.9.0
appnope==0.1.0
argon2-cffi==20.1.0
astroid==2.4.2
attrs==19.3.0
Babel==2.8.0
backcall==0.2.0
bert-score==0.3.5
bleach==3.1.5
blis==0.2.4
boto==2.49.0
boto3==1.14.47
botocore==1.17.47
bpemb==0.3.2
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.2
chakin==0.0.8
chardet==3.0.4
click==7.1.2
cloudpickle==1.5.0
colorama==0.4.3
conllu==1.3.1
cycler==0.10.0
cymem==2.0.3
dataclasses==0.7
dataenforce==0.1.2
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.10
docutils==0.15.2
easy-rouge==0.2.2
editdistance==0.5.3
entrypoints==0.3
filelock==3.0.12
flair==0.6
flaky==3.7.0
Flask==1.1.2
Flask-Cors==3.0.8
ftfy==5.8
future==0.18.2
gensim==3.8.3
gevent==20.6.2
gitdb==4.0.5
GitPython==3.1.3
greenlet==0.4.16
h5py==2.10.0
hyperopt==0.2.4
idna==2.10
imagesize==1.2.0
importlib-metadata==1.7.0
iniconfig==1.0.1
ipykernel==5.3.0
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.4.2
itsdangerous==1.1.0
Janome==0.3.10
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
json5==0.9.5
jsonnet==0.16.0
jsonpickle==1.4.1
jsonschema==3.2.0
jupyter-client==6.1.5
jupyter-core==4.6.3
jupyter-kite==0.2.1a0
jupyter-tabnine==1.1.0
jupyterlab==2.1.5
jupyterlab-git==0.20.0
jupyterlab-github==2.0.0
jupyterlab-server==1.1.5
kiwisolver==1.2.0
konoha==4.6.1
langdetect==1.0.8
lazy-object-proxy==1.4.3
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.4.0
mpld3==0.3
murmurhash==1.0.2
nbconvert==5.6.1
nbdime==2.0.0
nbformat==5.0.7
networkx==2.4
nltk==3.5
notebook==6.0.3
numpy==1.19.0
numpydoc==1.1.0
overrides==3.0.0
packaging==20.4
pandas==1.0.1
pandocfilters==1.4.2
parsimonious==0.8.1
parso==0.7.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plac==0.9.6
plotly==4.8.2
pluggy==0.13.1
preshed==2.0.1
progressbar2==3.51.4
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.13.0
ptyprocess==0.6.0
py==1.9.0
pybind11==2.5.0
pycparser==2.20
Pygments==2.6.1
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.0.1
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.34.1
python-utils==2.4.0
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2020.1
pyzmq==19.0.1
regex==2020.7.14
requests==2.24.0
responses==0.11.0
retrying==1.3.3
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.0
seaborn==0.10.1
segtok==1.5.10
Send2Trash==1.5.0
sentencepiece==0.1.91
six==1.15.0
smart-open==2.1.0
smmap==3.0.4
snowballstemmer==2.0.0
spacy==2.1.9
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlitedict==1.6.0
sqlparse==0.3.1
srsly==1.0.2
statsmodels==0.11.1
tabulate==0.8.7
tensorboardX==2.1
terminado==0.8.3
testpath==0.4.4
thinc==7.0.8
threadpoolctl==2.1.0
tokenizers==0.8.0rc4
toml==0.10.1
torch==1.5.0
torchtext==0.5.0
torchvision==0.6.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
transformers==3.0.0
typed-ast==1.4.1
ujson==1.35
Unidecode==1.1.1
urllib3==1.25.10
wasabi==0.7.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
word2number==1.1
wrapt==1.12.1
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
